+++ docs/docs/release-notes.md
@@ -15,6 +15,7 @@ sidebar_position: 200
 - **[Postgresql]**: Postgresql plugin defining technical assets
 - **[Docs]**: Adding an architecture overview
 - **[UI]**: Remove audit logs page
+- **[Output ports]**: Rename the `access_type` value `public` to `unrestricted` for output ports.
 
 ### bugfixes